made changes and added logreg file
